Consolidated all of the individual functions that checked and assigned group membership into a single flexible function that can assign membership of any group.
